@@ -24,67 +24,68 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-eslint/schematics": "^16.1.1",
-    "@angular/animations": "^16.1.6",
-    "@angular/common": "^15.2.4",
-    "@angular/core": "^15.2.4",
-    "@angular/fire": "^7.5.0",
-    "@angular/forms": "^15.2.4",
-    "@angular/localize": "^15.2.4",
-    "@angular/platform-browser": "^15.2.4",
-    "@angular/platform-browser-dynamic": "^15.2.4",
-    "@angular/router": "^15.2.4",
+    "@angular/animations": "^16.2.6",
+    "@angular/common": "^16.2.6",
+    "@angular/core": "16",
+    "@angular/fire": "^7.6.1",
+    "@angular/forms": "^16.2.6",
+    "@angular/localize": "^16.2.6",
+    "@angular/platform-browser": "^16.2.6",
+    "@angular/platform-browser-dynamic": "^16.2.6",
+    "@angular/router": "^16.2.6",
     "@aws-crypto/sha256-js": "^5.0.0",
-    "@capacitor/app": "4.1.1",
-    "@capacitor/core": "4.6.1",
-    "@capacitor/haptics": "4.1.0",
-    "@capacitor/keyboard": "4.1.0",
+    "@capacitor/android": "^5.4.1",
+    "@capacitor/app": "^5.0.6",
+    "@capacitor/core": "^5.4.1",
+    "@capacitor/haptics": "^5.0.6",
+    "@capacitor/ios": "^5.4.1",
+    "@capacitor/keyboard": "^5.0.6",
     "@capacitor/network": "^5.0.6",
-    "@capacitor/status-bar": "4.1.1",
-    "@ionic/angular": "^7.0.0-rc.5",
+    "@capacitor/status-bar": "^5.0.6",
+    "@ionic/angular": "^7.4.1",
     "@ionic/angular-server": "^7.0.0-rc.5",
     "@ionic/storage": "^3.0.6",
     "@ionic/storage-angular": "^4.0.0-next.4",
-    "@progress/kendo-angular-barcodes": "^11.4.0",
-    "@progress/kendo-angular-buttons": "^13.2.0",
-    "@progress/kendo-angular-charts": "^11.4.0",
-    "@progress/kendo-angular-common": "11.4.0",
-    "@progress/kendo-angular-conversational-ui": "^11.4.0",
-    "@progress/kendo-angular-dateinputs": "11.4.0",
-    "@progress/kendo-angular-dialog": "11.4.0",
-    "@progress/kendo-angular-dropdowns": "11.4.0",
-    "@progress/kendo-angular-editor": "^11.4.0",
-    "@progress/kendo-angular-excel-export": "11.4.0",
-    "@progress/kendo-angular-filter": "^11.4.0",
-    "@progress/kendo-angular-gauges": "^11.4.0",
-    "@progress/kendo-angular-grid": "^11.4.0",
-    "@progress/kendo-angular-icons": "11.4.0",
-    "@progress/kendo-angular-indicators": "^11.4.0",
-    "@progress/kendo-angular-inputs": "11.4.0",
-    "@progress/kendo-angular-intl": "11.4.0",
-    "@progress/kendo-angular-l10n": "11.4.0",
-    "@progress/kendo-angular-label": "11.4.0",
-    "@progress/kendo-angular-layout": "11.4.0",
-    "@progress/kendo-angular-listbox": "^11.4.0",
-    "@progress/kendo-angular-listview": "^11.4.0",
-    "@progress/kendo-angular-menu": "^11.4.0",
-    "@progress/kendo-angular-navigation": "11.4.0",
+    "@progress/kendo-angular-barcodes": "^13.5.0",
+    "@progress/kendo-angular-buttons": "^13.5.0",
+    "@progress/kendo-angular-charts": "^13.5.0",
+    "@progress/kendo-angular-common": "^13.5.0",
+    "@progress/kendo-angular-dateinputs": "^13.5.0",
+    "@progress/kendo-angular-dialog": "^13.5.0",
+    "@progress/kendo-angular-dropdowns": "^13.5.0",
+    "@progress/kendo-angular-editor": "^13.5.0",
+    "@progress/kendo-angular-excel-export": "^13.5.0",
+    "@progress/kendo-angular-filter": "^13.5.0",
+    "@progress/kendo-angular-gauges": "^13.5.0",
+    "@progress/kendo-angular-grid": "^13.5.0",
+    "@progress/kendo-angular-icons": "^13.5.0",
+    "@progress/kendo-angular-indicators": "^13.5.0",
+    "@progress/kendo-angular-inputs": "^13.5.0",
+    "@progress/kendo-angular-intl": "^13.5.0",
+    "@progress/kendo-angular-l10n": "^13.5.0",
+    "@progress/kendo-angular-label": "^13.5.0",
+    "@progress/kendo-angular-layout": "^13.5.0",
+    "@progress/kendo-angular-listbox": "^13.5.0",
+    "@progress/kendo-angular-listview": "^13.5.0",
+    "@progress/kendo-angular-menu": "^13.5.0",
+    "@progress/kendo-angular-navigation": "^13.5.0",
     "@progress/kendo-angular-pager": "^13.2.0",
-    "@progress/kendo-angular-pdf-export": "11.4.0",
-    "@progress/kendo-angular-popup": "11.4.0",
-    "@progress/kendo-angular-progressbar": "11.4.0",
-    "@progress/kendo-angular-scheduler": "^11.4.0",
-    "@progress/kendo-angular-scrollview": "^11.4.0",
-    "@progress/kendo-angular-toolbar": "11.4.0",
-    "@progress/kendo-angular-tooltip": "^11.4.0",
-    "@progress/kendo-angular-treeview": "11.4.0",
-    "@progress/kendo-angular-upload": "11.4.0",
-    "@progress/kendo-angular-utils": "^11.4.0",
+    "@progress/kendo-angular-pdf-export": "^13.5.0",
+    "@progress/kendo-angular-popup": "^13.5.0",
+    "@progress/kendo-angular-progressbar": "^13.5.0",
+    "@progress/kendo-angular-scheduler": "^13.5.0",
+    "@progress/kendo-angular-schematics": "^13.5.0",
+    "@progress/kendo-angular-scrollview": "^13.5.0",
+    "@progress/kendo-angular-toolbar": "^13.5.0",
+    "@progress/kendo-angular-tooltip": "^13.5.0",
+    "@progress/kendo-angular-treeview": "^13.5.0",
+    "@progress/kendo-angular-upload": "^13.5.0",
+    "@progress/kendo-angular-utils": "^13.5.0",
     "@progress/kendo-data-query": "^1.0.0",
     "@progress/kendo-date-math": "^1.5.10",
     "@progress/kendo-drawing": "^1.17.2",
     "@progress/kendo-file-saver": "^1.1.1",
-    "@progress/kendo-licensing": "^1.0.2",
+    "@progress/kendo-licensing": "^1.3.2",
     "@progress/kendo-svg-icons": "^1.0.0",
     "@progress/kendo-theme-default": "^6.0.2",
     "@types/crypto-js": "^4.1.1",
@@ -131,30 +132,30 @@
     "rxjs": "^7.5.0",
     "tslib": "^2.3.0",
     "yarn": "^1.22.19",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^15.0.0",
-    "@angular-devkit/build-angular": "^15.2.4",
+    "@angular-builders/custom-webpack": "^16.0.1",
+    "@angular-devkit/build-angular": "^16.2.3",
     "@angular-eslint/builder": "^14.0.0",
     "@angular-eslint/eslint-plugin": "^14.0.0",
     "@angular-eslint/eslint-plugin-template": "^14.0.0",
     "@angular-eslint/template-parser": "^14.0.0",
-    "@angular/cli": "^15.2.9",
-    "@angular/compiler": "^15.2.9",
-    "@angular/compiler-cli": "^15.2.4",
+    "@angular/cli": "^16.2.3",
+    "@angular/compiler": "^16.2.6",
+    "@angular/compiler-cli": "^16.2.6",
     "@angular/language-service": "^15.2.4",
-    "@capacitor/cli": "4.6.1",
+    "@capacitor/cli": "^4.6.1",
     "@ionic/angular-toolkit": "10.0.0",
     "@types/hammerjs": "^2.0.41",
     "@types/jasmine": "~4.0.0",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/parser": "^5.62.0",
+    "@typescript-eslint/parser": "^6.7.3",
     "babel-loader": "^9.1.2",
     "cordova-sqlite-storage": "^6.1.0",
-    "eslint": "^8.7.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsdoc": "30.7.6",
+    "eslint": "^8.50.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jsdoc": "^46.8.2",
     "eslint-plugin-prefer-arrow": "1.2.2",
     "jasmine-core": "~4.3.0",
     "jasmine-spec-reporter": "~5.0.0",
