@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
-
+import { loginGuard } from './guard/login-gurad/login-guard.guard';
+import { posGuard } from './guard/pos-gurad/pos-gurad.guard';
 const routes: Routes = [
   {
     path: '',
@@ -11,18 +12,24 @@ const routes: Routes = [
     path: 'system',
     loadChildren: () => import('./page/system/system.module').then(m => m.SystemPageModule),
   },
+
   {
     path: 'no-internet',
     loadChildren: () =>
       import('./page/access/no-internet/no-internet.module').then(m => m.NoInternetPageModule),
   },
   {
-    path: 'login',
-    loadChildren: () => import('./page/access/login/login.module').then(m => m.LoginPageModule),
-  },
-  {
     path: 'booking',
     loadChildren: () => import('./page/booking/booking.module').then(m => m.BookingPageModule),
+  },
+  {
+    path: 'login',
+    loadChildren: () => import('./page/access/login/login.module').then(m => m.LoginPageModule),
+    canActivate: [loginGuard],
+  },
+  {
+    path: 'shop',
+    loadChildren: () => import('./page/shop/shop.module').then(m => m.ShopPageModule),
   },
 ];
 
