@@ -1,7 +1,7 @@
-import { ShopCategoryType } from '../..';
+import { BookingSearchIconType } from '../..';
 import { PostCodeItemType } from '../../global/global.interface';
 
 export type BookingQueryCriteria = {
   address: PostCodeItemType;
-  category: ShopCategoryType;
+  shopCategoryName: BookingSearchIconType;
 };
