@@ -9,7 +9,7 @@ import { LoginPageRoutingModule } from './login-routing.module';
 import { LoginPage } from './login.page';
 import { LoginnModule } from 'src/app/component/user/login/login.module';
 import { AngularFireAuthModule } from '@angular/fire/compat/auth';
-import { BookingLanguageSelectionModule } from 'src/app/component/booking/booking-header/booking-language/booking-language-selection.module';
+import { BookingLanguageSelectionModule } from 'src/app/component/booking/booking/booking-header/booking-language/booking-language-selection.module';
 import { LanguageTransformPipeModule } from 'src/app/pipe/language-transform-pipe/language-transform.pipe.module';
 @NgModule({
   imports: [
